requirements.txt
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
binaryornot==0.4.4
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
convertdate==2.4.0
cookiecutter==2.6.0
cryptography==44.0.3
diff-match-patch==20241021
Django==5.2
django-browser-reload==1.18.0
django-debug-toolbar==5.1.0
django-environ==0.12.0
django-extensions==4.1
django-fernet-encrypted-fields==0.3.0
django-filter==25.1
django-htmx==1.23.0
django-import-export==4.3.7
django-tailwind==4.0.1
django-template-partials==24.4
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
executing==2.2.0
factory_boy==3.3.3
Faker==37.1.0
icecream==2.1.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
neapolitan==24.8
packaging==25.0
pluggy==1.5.0
pycparser==2.22
Pygments==2.19.1
PyMeeus==0.5.12
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
urllib3==2.4.0
